# public/swagger/openapi.yaml
...
openapi: "3.0.0"
info:
  title: Http Backend GO
  description: Api Docs
  version: 2.0.0
servers:
  - url: /
paths:
...
  /api/v1/authentication/refresh_token:
    post:
      tags:
      - authentication
      operationId: Refresh token
      requestBody:
        content:
          application/json:
            schema:
              required:
              - refreshToken
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        default:
          description: no description
        "200":
          description: OK
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT